Bug #30287668 [WARNING] INNODB: A LONG SEMAPHORE WAIT
Problem
-------
This bug is w.r.t to spatial indexes
When we are trying to merge two non-leaf pages
of a spatial index the comparison function says
that two records in the non-leaf pages are equal
.There is no condition written in function
rtr_page_copy_rec_list_end_no_locks() to handle this and
it loops around this function constantly iterating
over the record .

Analysis and FIX
----------------

The non-leaf page of the spatial index only contains
the MBR data and the child node pointer(page number)
unlike btree non leaf node which also contains the
PK keys. The leaf page of the spatial index do contain
the Pk Keys

Bug bug#22027053 GIS: PURGE THREAD DIES IN RTREE CODE WITH SPATIAL DATA
introduced a fix in which the record in a non leaf page
of a spatial index is ordered by MBR and page number
so we must compare MBR+Page number to order the records

In cmp_rec_rec_with_match() function when we are trying
to match the fields in the record of a non-leaf record
we get the column type from the dictionary .The dictionary
has no information that column values in a non-leaf are
different than the leaf node ,so it gives the data type
of the PK field and the comparison function uses it to
compare the child node pointer which gives wrong results.

The fix is to set the datatype of second field in the
non-leaf node to DATA_SYS_CHILD,so the page numbers
can be compared with proper datatype.

#rb 23242, 23283
Approved by: Yasufumi Kinoshita <yasufumi.kinoshita@oracle.com>